Rename auth2 to auth
