+++ handlers/db_dev.go
@@ -10,7 +10,7 @@ import (
 	"time"
 
 	"github.com/mtlynch/screenjournal/v2"
-	"github.com/mtlynch/screenjournal/v2/auth2"
+	"github.com/mtlynch/screenjournal/v2/auth"
 	"github.com/mtlynch/screenjournal/v2/handlers/parse"
 
 	"github.com/mtlynch/screenjournal/v2/random"
@@ -179,11 +179,11 @@ func (s Server) getDB(r *http.Request) store.Store {
 	return sharedDBSettings.GetDB(dbToken(c.Value))
 }
 
-func (s Server) getAuthenticator(r *http.Request) auth2.Authenticator {
+func (s Server) getAuthenticator(r *http.Request) auth.Authenticator {
 	if !sharedDBSettings.IsSessionIsolationEnabled() {
 		return s.authenticator
 	}
-	return auth2.New(s.getDB(r))
+	return auth.New(s.getDB(r))
 }
 
 func dbPerSessionPost() http.HandlerFunc {
