+++ handlers/db_prod.go
@@ -5,7 +5,7 @@ package handlers
 import (
 	"net/http"
 
-	"github.com/mtlynch/screenjournal/v2/auth2"
+	"github.com/mtlynch/screenjournal/v2/auth"
 	"github.com/mtlynch/screenjournal/v2/store"
 )
 
@@ -17,6 +17,6 @@ func (s Server) getDB(*http.Request) store.Store {
 	return s.store
 }
 
-func (s Server) getAuthenticator(r *http.Request) auth2.Authenticator {
+func (s Server) getAuthenticator(r *http.Request) auth.Authenticator {
 	return s.authenticator
 }
