+++ handlers/server.go
@@ -4,7 +4,7 @@ import (
 	"github.com/gorilla/mux"
 
 	"github.com/mtlynch/screenjournal/v2/announce"
-	"github.com/mtlynch/screenjournal/v2/auth2"
+	"github.com/mtlynch/screenjournal/v2/auth"
 	"github.com/mtlynch/screenjournal/v2/handlers/sessions"
 	"github.com/mtlynch/screenjournal/v2/metadata"
 	"github.com/mtlynch/screenjournal/v2/store"
@@ -12,7 +12,7 @@ import (
 
 type Server struct {
 	router         *mux.Router
-	authenticator  auth2.Authenticator
+	authenticator  auth.Authenticator
 	announcer      announce.Announcer
 	sessionManager sessions.Manager
 	store          store.Store
@@ -26,7 +26,7 @@ func (s Server) Router() *mux.Router {
 
 // New creates a new server with all the state it needs to satisfy HTTP
 // requests.
-func New(authenticator auth2.Authenticator, announcer announce.Announcer, sessionManager sessions.Manager, store store.Store, metadataFinder metadata.Finder) Server {
+func New(authenticator auth.Authenticator, announcer announce.Announcer, sessionManager sessions.Manager, store store.Store, metadataFinder metadata.Finder) Server {
 	s := Server{
 		router:         mux.NewRouter(),
 		authenticator:  authenticator,
